Refactor add project controller
