+++ server/src/controllers/projects/deleteProject.ts
