+++ server/src/controllers/projects/getProjects.ts
